tarefo_ai/requirements.txt
crewai
langchain
python-dotenv
pydantic
pillow
pyyaml
python-telegram-bot